Corrects point handling to match previous commit
The previous commit changed how point information is stored in the
annotation file: now each point is stored in the form [[x, y]].
This faciliates the multi-object handling, since now we can just append
new points to the list containing the last point. It also simplifies
the segmentation command. FastSAMPrompt takes a list of points in that
exact format, so we just adjust that in this commit.